largefiles: make update with backup files in .hglf slightly less broken
Largefiles update would try to copy f to f.orig if there was a .hglf/f.orig .
That is in many many ways very very wrong, but it also caused an abort if f
didn't exist.

Now it only tries to copy f if it exists.

+++ hgext/largefiles/lfcommands.py
@@ -483,8 +483,8 @@ def _updatelfile(repo, lfdirstate, lfile):
     abslfile = repo.wjoin(lfile)
     absstandin = repo.wjoin(lfutil.standin(lfile))
     if os.path.exists(absstandin):
-        if os.path.exists(absstandin+'.orig'):
-            shutil.copyfile(abslfile, abslfile+'.orig')
+        if os.path.exists(absstandin + '.orig') and os.path.exists(abslfile):
+            shutil.copyfile(abslfile, abslfile + '.orig')
         expecthash = lfutil.readstandin(repo, lfile)
         if (expecthash != '' and
             (not os.path.exists(abslfile) or
